Move TaskSignal to runtime module
